Rename latitude and longitude
