+++ .gitignore
@@ -1,1 +1,5 @@
 __pycache__
+
+dev_test.py
+
+*.pyc
